% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_model.R
\name{get_model}
\alias{get_model}
\title{Get the correct model}
\usage{
get_model(x, method, ...)
}
\arguments{
\item{x}{data}

\item{method}{The type of analysis. Possible analysis: c('time_series')}
}
\description{
Get the correct model
}
\examples{
model <- get_model(lh, 'time_series')
}
\author{
Emiel Veersma
}
\seealso{
Other get_model: \code{\link{get_model.time_series}}
}
\keyword{vfmodels}

