% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_model.R
\name{get_model.time_series}
\alias{get_model.time_series}
\title{Get Time Series Model}
\usage{
get_model.time_series(x, lag = NULL, labels = NULL, family = gaussian,
  train = NULL)
}
\arguments{
\item{x}{a univariate time series}

\item{lag}{the number of lags to incorporate in the model}

\item{labels}{the dates of the series}
}
\description{
An algorithm to find the appropriate time series model. The algorithm performs a forward search on the formula.
As a starting point, a completely empty formula is chosen and every time a sum or product is added/removed. The local optimum is
found by minimizing the Bayesian information criterion-value (BIC-value).

Some improvements still need to be made:
\itemize{
\item Adjust for heterskedasticity
\item Implement smoothing (now only sums and products of constants, linear relations and lags are incorporated)
\item Check for overfitting
}
}
\examples{
model <- get_model(lh, 'time_series')
}
\author{
Emiel Veersma
}
\seealso{
\url{http://www.automaticstatistician.com/research.php} for the inspiration of this function.

Other get_model: \code{\link{get_model}}
}
\keyword{series}
\keyword{time}
\keyword{vfmodels}

