% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_model.R
\name{plot.model}
\alias{plot.model}
\title{Plot a model}
\usage{
\method{plot}{model}(model)
}
\arguments{
\item{model}{a model object}
}
\description{
Plot a model found by get_model
}
\examples{
data('EIA')
train <- 1:(nrow(data) - 5)
labels <- data$Date
x <- data$Value

model <- get_model(x = x,method = 'time_series', lag = 2, labels = labels, family = gaussian('log'), train = train )
plot(model)
}
\author{
Emiel Veersma
}
\keyword{vfmodels}

