% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_model.R
\name{predict.model}
\alias{predict.model}
\title{Predict a model}
\usage{
\method{predict}{model}(model, ...)
}
\description{
Predict a model found by get_model. The input depends
}
\examples{
data('EIA')
train <- 1:(nrow(data) - 5)
labels <- data$Date
x <- data$Value

model <- get_model(x = x,method = 'time_series', lag = 2, labels = labels, family = gaussian('log'), train = train )
res <- predict(model, x = x, n.ahead = 2, labels = labels, train = train)
}
\author{
Emiel Veersma
}
\keyword{vfmodels}

