% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GitHub-package.R
\docType{package}
\name{vfmodels.timeseries}
\alias{vfmodels.timeseries}
\alias{vfmodels.timeseries-package}
\title{Veneficus Models}
\description{
An example package for GitHub
}
\details{
This package gives the user an insight in how the models at Veneficus help making the life of a Data Scientist a lot easier. Currently the package only consists of the get_model function, which automatically creates a model which describes the data as good as possible. this package will be expanded when new blogs get published.
}
\author{
Emiel Veersma \email{emiel.veersma@veneficus.nl}
}
\keyword{GitHub}
\keyword{package}
\keyword{vfmodels}

